backend/requirements.txt
django==3.2
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field[phonenumbers]~=6.0.0
phonenumbers~=8.12.43
djangorestframework~=3.13.1
requests~=2.27.1
geopy~=2.2.0
rollbar~=0.16.2
psycopg2~=2.9.3
gunicorn~=20.1.0
